fix(ci): address attachment lint and formatting
