+++ apps/web/src/lib/file-format.ts
@@ -1,0 +1,8 @@
+/** Compact human-readable byte count (e.g. 1.2 kB, 3.4 MB). */
+export function formatFileSize(bytes: number): string {
+	if (bytes <= 0) return "";
+	if (bytes < 1024) return `${bytes} B`;
+	if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} kB`;
+	if (bytes < 1024 * 1024 * 1024) return `${(bytes / 1024 / 1024).toFixed(1)} MB`;
+	return `${(bytes / 1024 / 1024 / 1024).toFixed(2)} GB`;
+}
